// src/app/models/Response/Produccion/tb-empleado/Vw-Empleado.ts
export class VwEmpleado {
    idEmpleado: number = 0;
    nombrePersona: string = "";
    apellidoEmp: string = "";
    tipoPersona: string = "";
    tipoDocumento: string = "";
    numeroDocumento: string = "";
    telefono: string = "";
    codigoUbigeo: string = "";
    direccion: string = "";
    salario: number = 0;
    estado: boolean = true;
}